+++ CHANGELOG.md
@@ -93,7 +93,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   for collections.  Added support for adding DirectMembershipRules for users and
   devices using resource name or ResourceId.
 - Added additional schedule types for CMCollections schedules.
-
+- Updated CMSystemDiscovery to warn when exceeding max allowed hours or days or minutes
+  and when specifying under the minimum allowed minutes.
 
 ### Removed